Mass Refactor + Integration Tests
Did a bunch of refactoring to disambiguate some of the codebase.
The compute engine (wrapper component) interacts with the `DataStorageSystem` rather than directly with `InputSource`s and `OutputSource`s.
Also ran a formatting pass on the codebase. If any other files upset checkstyles, I will fix them in the next commit.