Compute Engine Integration Test (#24)
* Create ComputeEngineIntegrationTest.java

started integration test for compute engine

* Refactoring Data Management

Removed type arguments from InputSource, and updated IOSources to have similar format to InMemoryIO Test variants. This clears a lot of confusion about typing, and should aim to make the code more readable.

* Fixed Compute Wrapper

* Corrected Imports on OutputSource

* Updated Method Signature on InMemoryInput

`write()` now takes a String as an argument as opposed to having no argument.

* Working on ComputeEngineIntegrationTest

Wrote in the additional testing, though I am unsure on how to run the tests to verify that they work.

* Mass Refactor + Integration Tests

Did a bunch of refactoring to disambiguate some of the codebase.
The compute engine (wrapper component) interacts with the `DataStorageSystem` rather than directly with `InputSource`s and `OutputSource`s.
Also ran a formatting pass on the codebase. If any other files upset checkstyles, I will fix them in the next commit.

---------

Co-authored-by: Melataylor519 <melataylor519@gmail.com>