Refactor Complete (#28)
Assignments 1-4 are complete, with 5 being a bit patchy in spots.

6 and onwards are more-or-less absent in the midst of refactoring the project for a second time.

* Attempting to add CSV-Based Input as Valid

Known no-compile. Looking for help with already declared signatures and those that need changing.
CSV Input are useful for in-memory as well as over the wire for user-specified input not from a file.

* Replace Iterable<Integer> with Iterator<Integer>

File I/O from file to Iterator<Integer> was fine, but everything else relied on Iterable<Integer>, so I changed everything that relied on Iterable to Iterator to continue adding support for Comma Separated Value input.

* Remove NoFinalizerCheck From Checkstyles

The sole cause of checks no longer passing, as the finalize method must be overridden for the guaranteed closing of file handles when the DataStorageSystemImpl gets GC'd. I respect the No-Finalizer Check, but it doesn't work here.